:truck: Move i18n related files to src/i18n
