+++ src/app/[locale]/layout.tsx
@@ -2,8 +2,8 @@ import type { Metadata } from 'next';
 import React from 'react';
 import { dir } from 'i18next';
 import '../globals.css';
-import { i18nConfig } from '@/i18nConfig';
-import initTranslations from '@/i18n';
+import { i18nConfig } from '@/i18n/i18nConfig';
+import initTranslations from '@/i18n/i18n';
 import TranslationsProvider from '@/components/i18n/TranslationsProvider';
 import SideBar from '@/components/common/sidebar/SideBar';
 import ConditionalStyledBackground from '@/components/common/ConditionalStyledBackground';
