+++ src/app/[locale]/page.tsx
@@ -1,10 +1,9 @@
-import initTranslations from '@/i18n';
 import Link from 'next/link';
-import TranslationsProvider from '@/components/TranslationsProvider';
+import React from 'react';
+import initTranslations from '@/i18n/i18n';
+import TranslationsProvider from '@/components/i18n/TranslationsProvider';
 import ExampleClientComponent from '@/components/ExampleClientComponent';
-import LanguageChanger from '@/components/LanguageChanger';
 
-const i18nNamespaces = ['home', 'common'];
 interface Params {
   params: {
     locale: string;
@@ -12,6 +11,7 @@ interface Params {
 }
 
 export default async function Home({ params: { locale } }: Params) {
+  const i18nNamespaces = ['home', 'common'];
   const { t, resources } = await initTranslations(locale, i18nNamespaces);
   return (
     <TranslationsProvider
@@ -22,7 +22,6 @@ export default async function Home({ params: { locale } }: Params) {
       <div>
         <h1>{t('greeting')}</h1>
         <ExampleClientComponent />
-        <LanguageChanger />
         <Link className="text-blue-600" href="/about-us">
           {t('common:about_us')}
         </Link>
