+++ src/components/i18n/TranslationsProvider.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { I18nextProvider } from 'react-i18next';
-import initTranslations from '@/i18n';
+import initTranslations from '@/i18n/i18n';
 import { createInstance, Resource } from 'i18next';
 import { ReactNode } from 'react';
