+++ src/i18n/i18n.ts
@@ -1,7 +1,7 @@
 import { createInstance, i18n, Resource } from 'i18next';
 import { initReactI18next } from 'react-i18next/initReactI18next';
 import resourcesToBackend from 'i18next-resources-to-backend';
-import { i18nConfig } from '@/i18nConfig';
+import { i18nConfig } from '@/i18n/i18nConfig';
 
 export default async function initTranslations(
   locale: string,
@@ -17,7 +17,7 @@ export default async function initTranslations(
     i18nInstance.use(
       resourcesToBackend(
         (language, namespace) =>
-          import(`../locales/${language}/${namespace}.json`)
+          import(`../../locales/${language}/${namespace}.json`)
       )
     );
   }
