+++ src/middleware.ts
@@ -1,5 +1,5 @@
 import { i18nRouter } from 'next-i18n-router';
-import { i18nConfig } from './i18nConfig';
+import { i18nConfig } from './i18n/i18nConfig';
 import { NextRequest } from 'next/server';
 
 export function middleware(req: NextRequest) {
